# speed up `first` and `only` for various types