@mkitti 2023-11-25
While I think you are technically correct, there is too much going on here at once.

At the moment, you are suggesting three distinguishable changes.

1. Acceralate the current `only` API for particular types.
2. Introduce a new public `only` API using `Val`.
3. Simplify and refactor old methods of `only`.

I suggest focusing this initial pull reqest on objective #1, which might be straightforward enough to advance quickly. If I understand correctly, this part of the pull request should just make existing API faster without a change in behavior. Therefore, it could advance as a pure performance pull request.

The `Val` usage needs a refactor for public API, but might work well enough for private methods for the moment. This is part probably needs to go to a triage review.

The simplification is fine but I think is fully dividable from this pull request since that should be effectively a no-op.