@matthias314 2023-11-29
I have made the following changes:

* rebased the PR on top of master
* renamed the new two-argument version of `only` to `_only`
* the second argument to `_only` now is `first` or `iterate`
* added `WeakKeyDict` and `PersistentDict` to the types to which the new method applies. The other types are `Dict`, `IdDict` and `Set`.
* change to `first` for `AbstractArray` partly undone. It now only adds `@propagate_inbounds`. (The previous version led to hickups in testing.)

The new methods for `only` are still defined where the types are defined except for `IdDict` (which is defined before iterators.jl). Doing it otherwise seems like a massive change to Base.jl for me.