@LilithHafner 2023-11-29
The string `only(` does not occur in the BaseBenchmarks repo so this won't be super informative, but `first(` does occur at least a couple of times

@nanosoldier `runbenchmarks("array" || "collection")`