@nanosoldier 2023-11-29
[Your benchmark job](https://github.com/JuliaLang/julia/pull/52296#issuecomment-1831145445) has completed - successfully executed benchmarks. A full report can be found [here](https://github.com/JuliaCI/NanosoldierReports/blob/master/benchmark/by_hash/56e30c5/report.md).