@LilithHafner 2023-12-01
@nanosoldier `runbenchmarks("array" || "collection", vs=:master)`