@nanosoldier 2023-12-01
[Your benchmark job](https://github.com/JuliaLang/julia/pull/52296#issuecomment-1836295286) has completed - possible performance regressions were detected. A full report can be found [here](https://github.com/JuliaCI/NanosoldierReports/blob/master/benchmark/by_hash/89080c8_vs_aef528d/report.md).