@mkitti 2023-12-01
`["array", "index", ("sumcartesian", "1:100000")]` requires some investigation.