@LilithHafner 2023-12-04
That benchmark runs in O(1) time with a low constant factor so I'm guessing it's benchmarking noise.

```julia
julia> function perf_sumcartesian(A)
           s = zero(eltype(A))
           for I in CartesianIndices(size(A))
               val = Base.unsafe_getindex(A, I)
               s += val
           end
           return s
       end
perf_sumcartesian (generic function with 1 method)

julia> using Chairmarks

julia> @b 1:100000 perf_sumcartesian
0.466 ns
```